pyproject.toml
[project]
name = "smart-recipe-finder"
version = "1.1.0"
description = " RESTful API for storing, searching, and retrieving recipes with advanced filtering and natural language search capabilities. Built with Python and FastAPI."
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "alembic>=1.17.2",
    "asyncpg>=0.31.0",
    "chromadb>=1.3.5",
    "cryptography>=46.0.3",
    "fastapi>=0.124.0",
    "inflect>=7.5.0",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.12.0",
    "sentence-transformers>=5.1.2",
    "sqlalchemy>=2.0.44",
    "sqlalchemy-utils>=0.42.0",
    "uvicorn[standard]>=0.38.0",
    "torch>=2.9.1",
    "boto3>=1.42.18",
    "python-multipart>=0.0.21",
    "pillow>=12.0.0",
    "python-magic>=0.4.27",
    "boto3-stubs[s3]>=1.42.18",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-dotenv>=0.5.2",
    "matplotlib>=3.10.7",
    "pre-commit>=4.5.1",
    "ruff>=0.14.10",
    "mypy>=1.19.1",
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.uv.sources]
torch = [
  { index = "pytorch-cpu", marker = "sys_platform == 'linux'" },
]

# pyproject.toml

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
# E, W - pycodestyle standards
# F - Pyflakes
# I - Isort
# B - Bugbear
select = ["E", "W", "F", "I", "B"]
ignore = []

[tool.ruff.format]
quote-style = "double"

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true
explicit_package_bases = true
allow_redefinition = true
plugins = [
    "pydantic.mypy",
    "sqlalchemy.ext.mypy.plugin"
]

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true
